+++ src/cue/maestrocuehandler.h
@@ -23,7 +23,7 @@ namespace PixelMaestro {
 
 			MaestroCueHandler(CueController* controller) : CueHandler(controller) { }
 			~MaestroCueHandler();
-			uint8_t* set_show(Event* events = nullptr, uint16_t num_events = 0, bool preserve_event_index = true);
+			uint8_t* set_show();
 			uint8_t* set_timer(uint16_t interval);
 			uint8_t* start();
 			uint8_t* stop();