+++ src/cue/showcuehandler.cpp
@@ -49,6 +49,9 @@ namespace PixelMaestro {
 
 	void ShowCueHandler::run(uint8_t *cue) {
 		Show* show = controller_->get_maestro()->get_show();
+
+		if (show == nullptr) return;
+
 		switch((Action)cue[Byte::ActionByte]) {
 			case Action::SetEvents:
 				{