+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is loosely based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - Animations
 	- Added new Fire animation.
+- Cues
+	- Removed Events list parameter from `MaestroCueHandler::set_show()`. You will need to call `ShowCueHandler::set_events()` separately to initialize the Event list.
 
 ## [v0.12] - 2018-01-22
 ### Changed